+++ client/components/Closet/index.js
@@ -1,1 +1,149 @@
-export {default as ClosetContainer} from './ClosetContainer'
+import React, {Component, useState} from 'react'
+import {connect} from 'react-redux'
+import {Link} from 'react-router-dom'
+import {getDresses, addWear} from '../../store/closet'
+import {setIsModalOpen} from '../../store/utils'
+import RemoveConfirmationModal from '../RemoveConfirmation'
+import store from '../../store'
+import axios from 'axios'
+import './index.scss'
+
+const baseClass = 'closet-container'
+
+class ClosetContainer extends Component {
+  constructor(props) {
+    super(props)
+    this.state = store.getState()
+    this.filterCostMoreThan50 = this.filterCostMoreThan50.bind(this)
+    this.sortByCost = this.sortByCost.bind(this)
+  }
+
+  async addWear(dressId) {
+    store.dispatch(addWear(dressId))
+    const dressToIncrement = await this.state.closet.dresses.filter(
+      dress => dress.id === dressId
+    )
+    dressToIncrement[0].wearCount += 1
+  }
+
+  async handleAccept(dressId) {
+    store.dispatch(setIsModalOpen(false))
+    await axios.delete(`/api/closet/${dressId}`)
+  }
+  showModal(dressId) {
+    store.dispatch(setIsModalOpen(true))
+  }
+
+  handleCancel() {
+    store.dispatch(setIsModalOpen(false))
+  }
+
+  async sortByCost() {
+    const sortedDresses = await this.state.closet.dresses.sort(function(a, b) {
+      return a.cost - b.cost
+    })
+  }
+
+  async filterCostMoreThan50() {
+    const dressesFiltered = await this.state.closet.dresses.filter(
+      dress => dress.cost > 50
+    )
+  }
+  componentDidMount() {
+    store.dispatch(getDresses())
+    this.unsubscribe = store.subscribe(() => this.setState(store.getState))
+  }
+
+  componentWillUnmount() {
+    this.unsubscribe()
+  }
+
+  render() {
+    return (
+      <div className={`${baseClass}`}>
+        <div className={`${baseClass}_title`}>micaela's closet</div>
+
+        <div className={`${baseClass}__menu`}>
+          <div
+            onClick={this.filterCostMoreThan50}
+            className={`${baseClass}__menu__button`}
+          >
+            filter
+          </div>
+
+          <div
+            onClick={this.sortByCost}
+            className={`${baseClass}__menu__button`}
+          >
+            sort
+          </div>
+
+          <Link to="/add">
+            <div className={`${baseClass}__menu__button`}>add an item </div>
+          </Link>
+        </div>
+        <div className={`${baseClass}_table`}>
+          {this.state.closet.dresses
+            ? this.state.closet.dresses.map((dress, i) => (
+                <div className={`${baseClass}_item`} key={i}>
+                  <Link to={`/closet/${dress.id}`}>
+                    <div className={`${baseClass}__item__wearcount`}>
+                      {dress.wearCount}
+                    </div>
+                    <img
+                      className={`${baseClass}_image-body`}
+                      src={dress.imageURL}
+                    />
+                    <div className={`${baseClass}_name-body`}>{dress.name}</div>
+                  </Link>
+                  <div className={`${baseClass}__buttons`}>
+                    <Link
+                      className={`${baseClass}__buttons__edit`}
+                      to={{pathname: `/closet/${dress.id}/edit`, state: dress}}
+                    >
+                      edit
+                    </Link>
+                    <div
+                      className={`${baseClass}__buttons__add`}
+                      onClick={() => this.addWear(dress.id)}
+                    >
+                      {' '}
+                      + wear{' '}
+                    </div>
+                    <div
+                      className={`${baseClass}__info__buttons__remove`}
+                      onClick={() => this.showModal(dress.id)}
+                    >
+                      remove
+                    </div>
+                  </div>
+                  {this.state.utils.isModalOpen ? (
+                    <RemoveConfirmationModal
+                      handleCancel={this.handleCancel}
+                      handleAccept={this.handleAccept}
+                      dressId={dress.id}
+                    />
+                  ) : null}
+                </div>
+              ))
+            : null}
+        </div>
+      </div>
+    )
+  }
+}
+
+const mapState = state => {
+  return {
+    dresses: state.closet.dresses
+  }
+}
+
+const mapDispatchToProps = dispatch => {
+  return {
+    getDresses: () => dispatch(getDresses()),
+    addWear: dressId => dispatch(addWear(dressId))
+  }
+}
+
+export default connect(mapState, mapDispatchToProps)(ClosetContainer)
