--- client/components/Search/SearchContainer.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react'
-
-class SearchContainer extends Component {
-  render() {
-    return (
-      <div>
-        <h1>search</h1>
-      </div>
-    )
-  }
-}
-
-export default SearchContainer
