+++ client/components/Search/index.js
@@ -1,1 +1,13 @@
-export {default as SearchContainer} from './SearchContainer'
+import React, {Component} from 'react'
+
+class Search extends Component {
+  render() {
+    return (
+      <div>
+        <h1>search</h1>
+      </div>
+    )
+  }
+}
+
+export default Search
